+++ test/services/database-analyzer/database-analyzer-mongo.js
@@ -15,7 +15,7 @@ const expectedManyuNullsModel = require('../../expected/mongo/db-analysis-output
 const expectedManyObjectIDFieldsModel = require('../../expected/mongo/db-analysis-output/many-objectid-fields.json');
 
 describe('Database analyser > MongoDB', () => {
-  describeMongoDatabases(mongoUrl => () => {
+  describeMongoDatabases((mongoUrl) => () => {
     let mongoHelper;
     let databaseConnection;