+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Technical - Replace `package-lock.json` by `yarn.lock`.
 - Technical - Use Yarn instead of NPM in the CI.
 - Technical - Rename `.env.example` file.
+- Technical - Upgrade ESLint rules.
 
 ### Fixed
 - MySQL - Handle BIT(1) boolean columns and handle the buffer<01> value returned on join queries.